More major fixes

+++ core/src/io/anuke/koru/network/packets/AnimationPacket.java
@@ -3,6 +3,6 @@ package io.anuke.koru.network.packets;
 import io.anuke.koru.renderers.AnimationType;
 
 public class AnimationPacket{
-	public long player;
+	public int player;
 	public AnimationType type;
 }
